% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bubble_plot.R
\name{Bubble_plot}
\alias{Bubble_plot}
\title{Create bubble plots}
\usage{
Bubble_plot(
  tabla_de_datos,
  tabla_de_datos_x,
  tabla_de_datos_y,
  tabla_de_datos_sz,
  tabla_de_datos_color
)
}
\arguments{
\item{tabla_de_datos}{a four row dataframe.}

\item{tabla_de_datos_x}{a valid col name to use on the x axis.}

\item{tabla_de_datos_y}{a valid col name to use on the y axis.}

\item{tabla_de_datos_sz}{a valid col name to use to plot the size of the bubbles.}

\item{tabla_de_datos_color}{a valid col name to use to plot the color of the bubbles.}
}
\description{
This function takes a four row dataframe to create a bubble plot.
}
\details{
This function is part of a package used for the analysis of microbial metabolism.
}
\examples{
Bubble_plot(example,  "gene", "genomes", "abundance", "group")
}
