% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ko_pheatmap_for_CP9_groups.R
\name{ko_pheatmap_for_CP9_groups}
\alias{ko_pheatmap_for_CP9_groups}
\title{Create a heatmap.}
\usage{
ko_pheatmap_for_CP9_groups(Table_groups, Key_vector, Table_genes, Order_vector)
}
\arguments{
\item{Table_groups}{a dataframe containing the information of the ko and groups.}

\item{Key_vector}{a vector with the interesting KO.}

\item{Table_genes}{a dataframe containing the names of the KO and the Genes.}

\item{Order_vector}{vector containing the order of the Genes in the plot.}
}
\description{
This function creates a heatmap of the genes in which you are interested.
}
\details{
This function is part of a package used for the analysis of microbial metabolism.
}
\examples{
ko_pheatmap_for_CP9_groups(Ko_Final_groups, Arsenic,  Input_data, orden_gene)
}
